lsp-devtools: move old code to temp module
